refactor: move factories, bench, client into test/
