+++ src/storage/db/reaction.test.ts
@@ -1,5 +1,5 @@
 import ReactionDB from '~/storage/db/reaction';
-import { Factories } from '~/factories';
+import { Factories } from '~/test/factories';
 import { jestRocksDB } from '~/storage/db/jestUtils';
 import { Ed25519Signer, ReactionAdd, ReactionRemove } from '~/types';
 import { NotFoundError } from '~/errors';
