+++ src/storage/engine/engine.verification.test.ts
@@ -1,5 +1,5 @@
 import Engine from '~/storage/engine';
-import { Factories } from '~/factories';
+import { Factories } from '~/test/factories';
 import {
   Ed25519Signer,
   EthereumSigner,
